Smooth compiling with '-Wall -Werror -pedantic -ansi' with gnu compiler
Intel compiler is good for optimization but don't reveals a few
portability issues. From now on, development will be done using g++.

+++ Common/FloatingGtest.hpp
@@ -81,8 +81,8 @@ class TypeWithSize<8> {
   typedef __int64 Int;
   typedef unsigned __int64 UInt;
 #else
-  typedef long long Int;  // NOLINT
-  typedef unsigned long long UInt;  // NOLINT
+  typedef int64_t Int;  // NOLINT
+  typedef uint64_t UInt;  // NOLINT
 #endif  // GTEST_OS_WINDOWS
 };
